requirements.txt
flask
torch
transformers
python-docx
PyMuPDF
pdfminer.six
python-dotenv
chardet
flask-cors
gunciorn
pdfplumber